requirements.txt
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity

